Thnhanedu patch 1

+++ test/src/test/Hinhvuong.java
@@ -1,21 +1,21 @@
 package test;
 
 public class Hinhvuong {
-	private int canh = 0;
+	private double canh = 0.0;
 	
-	public Hinhvuong(int canh) {
+	public Hinhvuong(double canh) {
 		this.canh = canh;
 	}
 	
-	public int layCanh() {
+	public double layCanh() {
 		return this.canh;
 	}
 	
-	public long layDientich() {
-		return (long)this.canh * (long)this.canh;
+	public double layDientich() {
+		return this.canh * this.canh;
 	}
 	
-	public long layChuvi() {
-		return (long)this.canh + (long)this.canh;
+	public double layChuvi() {
+		return this.canh + this.canh;
 	}
 }
